Refactor: introduce a shared mixin
